move json ser/deser of message and delivery to protocol module
