+++ src/protocol/mod.rs
@@ -332,6 +332,58 @@ impl Message {
     pub fn task_id(&self) -> &str {
         &self.headers.id
     }
+
+    pub fn json_serialized(&self) -> Result<Vec<u8>, serde_json::error::Error>{
+        use serde_json::{json,value::Value};
+        
+        let root_id = match &self.headers.root_id{
+            Some(root_id) => json!(root_id.clone()),
+            None => Value::Null
+        };
+        let reply_to = match &self.properties.reply_to{
+            Some(reply_to) => json!(reply_to.clone()),
+            None => Value::Null
+        };
+        let eta = match self.headers.eta{
+            Some(time) => json!(time.to_rfc3339()),
+            None => Value::Null
+        };
+        let expires = match self.headers.expires{
+            Some(time) => json!(time.to_rfc3339()),
+            None => Value::Null
+        };
+        let mut buffer = Uuid::encode_buffer();
+        let uuid = Uuid::new_v4().to_hyphenated().encode_lower(&mut buffer);
+        let delivery_tag = uuid.to_owned();
+        let msg_json_value = json!({
+            "body": self.raw_body.clone(),
+            "content_encoding": self.properties.content_encoding.clone(),
+            "content_type": self.properties.content_type.clone(),
+            "correlation_id": self.properties.correlation_id.clone(),
+            "reply_to": reply_to,
+            "delivery_tag": delivery_tag,
+            "headers": {
+                "id": self.headers.id.clone(),
+                "task": self.headers.task.clone(),
+                "lang": self.headers.lang.clone(),
+                "root_id": root_id,
+                "parent_id": self.headers.parent_id.clone(),
+                "group": self.headers.group.clone(),
+                "meth": self.headers.meth.clone(),
+                "shadow": self.headers.shadow.clone(),
+                "eta": eta,
+                "expires": expires,
+                "retries": self.headers.retries.clone(),
+                "timelimit": self.headers.timelimit.clone(),
+                "argsrepr": self.headers.argsrepr.clone(),
+                "kwargsrepr": self.headers.kwargsrepr.clone(),
+                "origin": self.headers.origin.clone()
+            }
+        });
+        let res = serde_json::to_string(&msg_json_value)?;
+        Ok(res.into_bytes())
+        // Ok(res.bytes())
+    }
 }
 
 impl<T> TryFrom<Signature<T>> for Message
@@ -524,6 +576,71 @@ pub struct MessageBodyEmbed {
     /// The serialized signature of the chord callback.
     #[serde(default)]
     pub chord: Option<String>,
+}
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct DeliveryHeaders{
+    pub id: String,
+    pub task: String,
+    pub lang: Option<String>,
+    pub root_id: Option<String>,
+    pub parent_id: Option<String>,
+    pub group: Option<String>,
+    pub meth: Option<String>,
+    pub shadow: Option<String>,
+    pub eta: Option<DateTime<Utc>>,
+    pub expires: Option<DateTime<Utc>>,
+    pub retries: Option<u32>,
+    pub timelimit: (Option<u32>, Option<u32>),
+    pub argsrepr: Option<String>,
+    pub kwargsrepr: Option<String>,
+    pub origin: Option<String>,
+}
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct Delivery{
+    pub body: Vec<u8>,
+    pub content_encoding: String,
+    pub content_type: String,
+    pub correlation_id: String,
+    pub reply_to: Option<String>,
+    pub delivery_tag: String,
+    pub headers: DeliveryHeaders,
+}
+
+
+impl Delivery{
+    pub fn try_deserialize_message(&self) -> Result<Message, ProtocolError>{
+        Ok(Message {
+            properties: MessageProperties {
+                correlation_id: self
+                    .correlation_id.clone(),
+                content_type: self
+                    .content_type.clone(),
+                content_encoding: self
+                    .content_encoding.clone(),
+                reply_to: self.reply_to.clone(),
+            },
+            headers: MessageHeaders {
+                id: self.headers.id.clone(),
+                task: self.headers.task.clone(),
+                lang: self.headers.lang.clone(),
+                root_id: self.headers.root_id.clone(),
+                parent_id: self.headers.parent_id.clone(),
+                group: self.headers.group.clone(),
+                meth: self.headers.meth.clone(),
+                shadow: self.headers.shadow.clone(),
+                eta: self.headers.eta.clone(),
+                expires: self.headers.expires.clone(),
+                retries: self.headers.retries.clone(),
+                timelimit: self.headers.timelimit.clone(),
+                argsrepr: self.headers.argsrepr.clone(),
+                kwargsrepr: self.headers.kwargsrepr.clone(),
+                origin: self.headers.origin.clone(),
+            },
+            raw_body: self.body.clone(),
+        })
+    }
 }
 
 #[cfg(test)]
